refactor cmd dir
